Organize http tests
