+++ test/integration-tests.ts
@@ -479,58 +479,4 @@ describe("Integration tests", () => {
     let actual: string = await clientOverride.getValueAsync("stringDefaultCat", defaultValue);
     assert.strictEqual(actual, "NOT_CAT");
   });
-
-  it("HTTP timeout", async () => {
-    const mockttp = require('mockttp');
-    const server = mockttp.getLocal({
-      https: {
-        keyPath: './test/cert/testCA.key',
-        certPath: './test/cert/testCA.pem'
-      }
-    });
-    server.forAnyRequest().thenTimeout();
-    await server.start();
-
-    const client = configcatClient.createClientWithManualPoll(sdkKey, {
-      requestTimeoutMs: 1000,
-      baseUrl: server.url
-    });
-    const startTime = new Date().getTime();
-    await client.forceRefreshAsync();
-    const duration = new Date().getTime() - startTime;
-    assert.isTrue(duration > 1000 && duration < 2000);
-
-    const defaultValue = "NOT_CAT"
-    assert.strictEqual(defaultValue, await client.getValueAsync("stringDefaultCat", defaultValue));
-
-    await server.stop();
-  });
-
-  it("HTTP proxy", async () => {
-    const mockttp = require('mockttp');
-    const server = mockttp.getLocal({
-      https: {
-        keyPath: './test/cert/testCA.key',
-        certPath: './test/cert/testCA.pem'
-      }
-    });
-    let proxyCalled = false;
-    server.forAnyRequest().forHost("cdn-global.configcat.com").thenPassThrough({
-      beforeRequest: (_: any) => {
-        proxyCalled = true;
-      }
-    });
-    await server.start();
-
-    const client = configcatClient.createClientWithManualPoll(sdkKey, {
-      proxy: server.url
-    });
-    await client.forceRefreshAsync();
-    assert.isTrue(proxyCalled);
-
-    const defaultValue = "NOT_CAT"
-    assert.strictEqual("Cat", await client.getValueAsync("stringDefaultCat", defaultValue));
-
-    await server.stop();
-  });
 });
